requirements.txt
streamlit
pandas
numpy
google-cloud-bigquery
google-auth
python-dotenv
streamlit-extras